refactor(dt-eval-lib): move deterministic param types into engine/deterministic
Relocate ExactMatchParams / RegexParams / ContainsParams / JsonSchemaParams
and DeterministicParams out of prompts/types.ts into
engine/deterministic/types.ts, next to the scorers and DeterministicOutcome,
so prompts/types.ts is not cluttered with method-specific params.
PromptDefinition imports DeterministicParams from the new location; the public
export surface (src/index.ts) is unchanged.

Also fixes the must_contain/must_not_contain empty-keywords error to actually
interpolate the method name (backticks instead of a plain string).
