+++ dt-eval-lib/src/prompts/index.ts
@@ -2,13 +2,5 @@ export { createCustomPrompt, deleteCustomPrompt } from "./custom";
 export { getPrompt, listPrompts } from "./lookup";
 export type { PromptStore } from "./store";
 export { registerPromptStore } from "./store";
-export type {
-  ContainsParams,
-  DeterministicParams,
-  EvaluatorMethod,
-  ExactMatchParams,
-  JsonSchemaParams,
-  PromptDefinition,
-  RegexParams,
-} from "./types";
+export type { EvaluatorMethod, PromptDefinition } from "./types";
 export { BuiltInMetric, DRIFT_METRIC_ID } from "./types";
