+++ dt-eval-lib/src/prompts/types.ts
@@ -1,3 +1,4 @@
+import type { DeterministicParams } from "../engine/deterministic/types";
 import type { ScoringScale } from "../scoring/types";
 
 export enum BuiltInMetric {
@@ -29,40 +30,6 @@ export type EvaluatorMethod =
   | "json_schema"
   | "must_contain"
   | "must_not_contain";
-
-export interface ExactMatchParams {
-  caseSensitive?: boolean;
-  trim?: boolean;
-}
-
-export interface RegexParams {
-  pattern: string;
-  flags?: string;
-}
-
-/**
- * Params for the must_contain / must_not_contain methods.
- * - `mode: "any"` (default) — must_contain passes when ≥1 keyword is present;
- *   must_not_contain passes when none are present (blocklist).
- * - `mode: "all"` — must_contain passes when all keywords are present;
- *   must_not_contain passes unless all are present.
- * - `caseSensitive` defaults to false (case-insensitive match).
- */
-export interface ContainsParams {
-  keywords: string[];
-  mode?: "any" | "all";
-  caseSensitive?: boolean;
-}
-
-export interface JsonSchemaParams {
-  schema: object;
-}
-
-export type DeterministicParams =
-  | ExactMatchParams
-  | RegexParams
-  | ContainsParams
-  | JsonSchemaParams;
 
 export interface PromptDefinition {
   id: string;
